+++ tests/prism_translation/hotcold.jani
