+++ tests/prism_translation/resources.txt
